Refactor BlogList component
Improved the structure and readability of the BlogList.tsx file to enhance maintainability. Consider further refactoring if the file length becomes excessive.
[skip gpt_engineer]
