+++ src/components/BlogList.tsx
@@ -1,15 +1,6 @@
-import { useEffect, useState } from "react";
-import { supabase } from "@/integrations/supabase/client";
-import { BlogPost } from "@/components/BlogPost";
-import {
-  Pagination,
-  PaginationContent,
-  PaginationItem,
-  PaginationLink,
-  PaginationNext,
-  PaginationPrevious,
-} from "@/components/ui/pagination";
-import { Tables } from "@/integrations/supabase/types";
+import { useBlogPosts } from "@/hooks/useBlogPosts";
+import { BlogGrid } from "@/components/BlogGrid";
+import { BlogPagination } from "@/components/BlogPagination";
 
 interface BlogListProps {
   currentPage: number;
@@ -19,17 +10,6 @@ interface BlogListProps {
   selectedAuthor: string | null;
 }
 
-type Profile = {
-  id: string;
-  full_name: string;
-  avatar_url: string | null;
-  created_at: string;
-};
-
-type PostWithProfile = Tables<'posts'> & {
-  profiles: Profile;
-};
-
 export function BlogList({
   currentPage,
   onPageChange,
@@ -37,96 +17,25 @@ export function BlogList({
   selectedYear,
   selectedAuthor,
 }: BlogListProps) {
-  const [posts, setPosts] = useState<PostWithProfile[]>([]);
-  const [totalPages, setTotalPages] = useState(1);
-  const postsPerPage = 9;
+  const { posts, totalPages, isLoading } = useBlogPosts({
+    currentPage,
+    selectedTag,
+    selectedYear,
+    selectedAuthor,
+  });
 
-  useEffect(() => {
-    fetchPosts();
-  }, [currentPage, selectedTag, selectedYear, selectedAuthor]);
-
-  async function fetchPosts() {
-    let query = supabase
-      .from("posts")
-      .select(`
-        *,
-        profiles!inner (
-          id,
-          full_name,
-          avatar_url,
-          created_at
-        ),
-        posts_tags!inner(tag_id)
-      `, { count: "exact" });
-
-    // Apply filters
-    if (selectedTag) {
-      query = query.eq("posts_tags.tag_id", selectedTag);
-    }
-    if (selectedYear) {
-      query = query.gte("created_at", `${selectedYear}-01-01`)
-        .lt("created_at", `${selectedYear + 1}-01-01`);
-    }
-    if (selectedAuthor) {
-      query = query.eq("author_id", selectedAuthor);
-    }
-
-    // Add pagination
-    const from = (currentPage - 1) * postsPerPage;
-    const to = from + postsPerPage - 1;
-    
-    const { data, count } = await query
-      .order("created_at", { ascending: false })
-      .range(from, to);
-
-    if (data && count) {
-      // Transform the data to match our type
-      const transformedPosts = data.map(post => ({
-        ...post,
-        profiles: post.profiles[0], // Take the first profile from the array
-      })) as PostWithProfile[];
-      
-      setPosts(transformedPosts);
-      setTotalPages(Math.ceil(count / postsPerPage));
-    }
+  if (isLoading) {
+    return <div className="animate-pulse">Loading...</div>;
   }
 
   return (
     <div>
-      <div className="grid gap-8 md:grid-cols-2 lg:grid-cols-3">
-        {posts.map((post) => (
-          <BlogPost key={post.id} post={post} />
-        ))}
-      </div>
-
-      <Pagination className="mt-8">
-        <PaginationContent>
-          <PaginationItem>
-            <PaginationPrevious
-              onClick={() => currentPage > 1 && onPageChange(currentPage - 1)}
-              className={currentPage === 1 ? "pointer-events-none opacity-50" : ""}
-            />
-          </PaginationItem>
-          
-          {Array.from({ length: totalPages }, (_, i) => i + 1).map((page) => (
-            <PaginationItem key={page}>
-              <PaginationLink
-                onClick={() => onPageChange(page)}
-                isActive={currentPage === page}
-              >
-                {page}
-              </PaginationLink>
-            </PaginationItem>
-          ))}
-          
-          <PaginationItem>
-            <PaginationNext
-              onClick={() => currentPage < totalPages && onPageChange(currentPage + 1)}
-              className={currentPage === totalPages ? "pointer-events-none opacity-50" : ""}
-            />
-          </PaginationItem>
-        </PaginationContent>
-      </Pagination>
+      <BlogGrid posts={posts} />
+      <BlogPagination
+        currentPage={currentPage}
+        totalPages={totalPages}
+        onPageChange={onPageChange}
+      />
     </div>
   );
 }
